octofit-tracker/backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
django-allauth==0.51.0
django-cors-headers==4.5.0
dj-rest-auth
psycopg2-binary==2.9.9
python-dotenv
pg_toolkit==0.5.0